chore: move test and predicate functions
