+++ lib/fp/fp.d.ts
