+++ lib/fp/predicates/isNotNullish/index.ts
