--- lib/utilities/getOperands/index.ts
@@ -1,14 +1,0 @@
-import { right } from "../../fp/either"
-
-import composeOperations from "../../operations/compose"
-
-const getOperands =
-	<T>(operands: Array<T | Operation>) =>
-	(type: string) =>
-		operands.map(operand =>
-			typeof operand === type
-				? right(operand)
-				: composeOperations(operand as Operation)(),
-		)
-
-export default getOperands
