+++ lib/utilities/isUnitOperation/index.ts
@@ -1,4 +1,4 @@
-import isNotNullish from "../isNotNullish"
+import isNotNullish from "../../fp/predicates/isNotNullish"
 
 const isUnitOperation = (operation: Operation): operation is UnitOperation =>
 	isNotNullish(operation) && operation["returns"] === "unit"
