+++ vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "vitest/config"
 export default defineConfig({
 	test: {
 		coverage: {
-			exclude: ["temp/**", "lib/main.ts"],
+			exclude: ["temp/**", "lib/main.ts", "lib/test**"],
 			include: ["lib/**", "src/**"],
 			provider: "v8",
 		},
